+++ src/components/Header/index.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+const Header: React.FC = () => {
+  return (
+    <header>
+      <span>Welcome to Awesome Movies</span>
+    </header>
+  );
+};
+
+export default Header;
