+++ src/components/LatestMovie/styles.css
@@ -1,0 +1,34 @@
+.movies-grid {
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+}
+
+.movies-grid li {
+  display: flex;
+  flex-direction: column;
+
+  list-style: none;
+  margin: 1rem;
+}
+
+.movies-grid li img {
+  cursor: pointer;
+
+  opacity: 0.7;
+  transition: 0.5s;
+}
+
+.movies-grid li img:hover {
+  opacity: 1;
+}
+
+.movies-grid li .movie-info strong {
+  font-size: 0.8rem;
+}
+
+.movies-grid li .movie-info span {
+  margin-left: 5px;
+
+  font-size: 0.8rem;
+}
