+++ src/components/Layout/styles.css
@@ -1,0 +1,14 @@
+#home-container {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+
+  height: 100vh;
+}
+
+header {
+  margin-top: 60px;
+
+  font-size: 2.5rem;
+  font-weight: 700;
+}
