+++ src/components/Main/styles.css
@@ -1,0 +1,8 @@
+#content {
+  max-width: 1100px;
+
+  margin: 60px;
+
+  display: flex;
+  flex-direction: column;
+}
