+++ src/components/NowPlayingMovie/styles.css
@@ -1,0 +1,56 @@
+.load-more {
+  font-size: 2.5rem;
+
+  padding: 0 auto;
+
+  cursor: pointer;
+
+  width: 100%;
+  height: 50px;
+  border-style: none;
+  outline: 0;
+
+  transition: 0.5s;
+
+  margin-bottom: 30px;
+}
+
+.load-more:hover {
+  opacity: 1;
+  background-color: aquamarine;
+}
+
+.movies-grid {
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+}
+
+.movies-grid li {
+  display: flex;
+  flex-direction: column;
+
+  list-style: none;
+  margin: 1rem;
+}
+
+.movies-grid li img {
+  cursor: pointer;
+
+  opacity: 0.7;
+  transition: 0.5s;
+}
+
+.movies-grid li img:hover {
+  opacity: 1;
+}
+
+.movies-grid li .movie-info strong {
+  font-size: 0.8rem;
+}
+
+.movies-grid li .movie-info span {
+  margin-left: 5px;
+
+  font-size: 0.8rem;
+}
