--- src/pages/Home/styles.css
@@ -1,83 +1,0 @@
-#home-container{
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-
-    height: 100vh;
-}
-
-header{
-    margin-top: 60px;
-
-    font-size: 2.5rem;
-    font-weight: 700;
-}
-
-#content{
-    max-width: 1100px;
-    
-    margin: 60px;
-}
-
-main{
-    display: flex;
-    flex-direction: column;
-}
-
-main .load-more{
-    font-size: 2.5rem;
-
-    padding: 0 auto;
-
-    cursor: pointer;
-
-    width: 100%;
-    height: 50px;
-    border-style: none;
-    outline: 0;
-
-    transition: 0.5s;
-
-    margin-bottom: 30px;
-}
-
-main .load-more:hover{
-    opacity: 1;
-    background-color: aquamarine;
-}
-
-main .movies-grid{
-    display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-}
-
-main .movies-grid li{
-    display: flex;
-    flex-direction: column;
-
-    list-style: none;
-    margin: 1rem;
-}
-
-main .movies-grid li img{
-    cursor: pointer;
-
-    opacity: 0.7;
-    transition: 0.5s;
-}
-
-main .movies-grid li img:hover{
-    opacity: 1;    
-}
-
-main .movies-grid li .movie-info strong{
-    font-size: 0.8rem;
-}
-
-main .movies-grid li .movie-info span{
-    margin-left: 5px;
-
-    font-size: 0.8rem;
-}
-
